backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["kvydyk <kvydyk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "24.1.0"
aiosmtplib = "2.0.2"
aiosqlite = "0.17.0"
alembic = "1.13.2"
amqp = "5.2.0"
annotated-types = "0.7.0"
anyio = "4.4.0"
asgiref = "3.8.1"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
babel = "2.15.0"
bcrypt = "4.1.3"
billiard = "4.2.0"
blinker = "1.8.2"
celery = "5.4.0"
certifi = "2024.7.4"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
cryptography = "42.0.8"
dnspython = "2.6.1"
email-validator = "2.2.0"
exceptiongroup = "1.2.2"
fastapi = "0.114.2"
fastapi-cli = "0.0.4"
fastapi-mail = "1.4.1"
flower = "2.0.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
humanize = "4.10.0"
idna = "3.7"
iso8601 = "1.1.0"
jinja2 = "3.1.4"
kombu = "5.3.7"
mako = "1.3.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
orjson = "3.10.6"
pendulum = "3.0.0"
phonenumbers = "8.13.40"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.47"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-extra-types = "2.9.0"
pydantic-settings = "2.3.4"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyjwt = "2.8.0"
pypika-tortoise = "0.1.6"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pytz = "2024.1"
pyyaml = "6.0.1"
redis = "5.0.7"
requests = "2.32.3"
rich = "13.7.1"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
sqladmin = "0.18.0"
sqlalchemy = "2.0.35"
starlette = "0.37.2"
time-machine = "2.14.2"
tornado = "6.4.1"
tortoise-orm = "0.21.5"
typer = "0.12.3"
typing-extensions = "4.12.2"
tzdata = "2024.1"
ujson = "5.10.0"
urllib3 = "2.2.3"
uvicorn = "0.30.1"
uvloop = "0.19.0"
vine = "5.1.0"
watchfiles = "0.22.0"
wcwidth = "0.2.13"
websockets = "12.0"
wtforms = "3.1.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
pytest-celery = "^1.1.3"
apscheduler = "^3.11.0"
factory-boy = "^3.3.1"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
